code review:move functions
